+++ packages/vue-dot/jest.config.ts
@@ -4,10 +4,7 @@ const config: Config.InitialOptions = {
 	preset: '@vue/cli-plugin-unit-jest/presets/typescript',
 	testMatch: [
 		'<rootDir>/src/**/tests/*.spec.ts',
-		'<rootDir>/tests/**/*.spec.ts'
-	],
-	testPathIgnorePatterns: [
-		'.e2e.spec.ts'
+		'<rootDir>/tests/unit/**/*.spec.ts'
 	],
 	moduleNameMapper: {
 		'^@/tests$': '<rootDir>/tests/index.ts',
@@ -27,9 +24,7 @@ const config: Config.InitialOptions = {
 		// for now because the report is incorrect
 		'src/**/*.ts',
 		'!**/node_modules/**',
-		'!**/*.e2e.spec.ts',
 		'!**/*.d.ts',
-		'!**/*.e2e.spec.ts',
 		'!**/dev/**',
 		'!**/data/**'
 	],