+++ packages/vue-dot/src/elements/CopyBtn/CopyBtn.vue
@@ -7,7 +7,7 @@
 		>
 			<template #activator="{ on }">
 				<VBtn
-					data-cy="btn"
+					data-cy="v-btn"
 					v-bind="options.btn"
 					:aria-label="label"
 					v-on="on"