+++ packages/vue-dot/tests/e2e/integration/CopyBtn.e2e.spec.ts
@@ -1,0 +1,15 @@
+import { mountComponent } from '..';
+// import CopyBtn from '@/elements/CopyBtn';
+
+describe('CopyBtn', () => {
+	const textToCopy = 'Texte copié !';
+
+	/**beforeEach(() => {
+		mountComponent(CopyBtn);
+	});
+*/
+	it('copies the text to the clipboard', () => {
+		//cy.dataCy('v-btn').should('exist').click();
+		// cy.document().contains(textToCopy);
+	});
+});