+++ packages/vue-dot/vue.config.js
@@ -1,6 +1,5 @@
 // Build configuration
 const webpack = require('webpack');
-var path = require('path');
 
 // If LIB_MODE is true, we're building the library
 // else, we're building the dev environment
@@ -76,7 +75,6 @@ const DEV_MODE_CONFIG = {
 
 				return args;
 			});
-		config.resolve.alias.set('@e2e/config', path.resolve(__dirname, './e2e/config/index.ts'));
 	},
 	transpileDependencies: [
 		'vuetify',